+++ internal/cmd/root.go
@@ -56,7 +56,7 @@ type CLI struct {
 	Auth         AuthCmd         `cmd:"" help:"Manage authentication"`
 	Accounts     AccountsCmd     `cmd:"" help:"Manage messaging accounts"`
 	Contacts     ContactsCmd     `cmd:"" help:"Search contacts"`
-	Assets       AssetsCmd       `cmd:"" help:"Download assets"`
+	Assets       AssetsCmd       `cmd:"" help:"Manage assets"`
 	Chats        ChatsCmd        `cmd:"" help:"Manage chats"`
 	Messages     MessagesCmd     `cmd:"" help:"Manage messages"`
 	Reminders    RemindersCmd    `cmd:"" help:"Manage chat reminders"`